+++ NEWS.md
@@ -3,8 +3,10 @@
 - updated `check_inventory()` to work with CORDEX-Adjust data, too. Check for full period also works if historical is merged to the RCP in filenames.
 - added `date_range` parameter to `nc_grid_to_dt()` to allow easier extraction of subset of data
 - added helper function `get_varnames()` to list variable names in a netcdf file 
+- `variable` argument in `nc_grid_to_dt()` and `rotpole_nc_point_to_dt()` will be guessed if not supplied 
 - made the inventory data.table from `get_inventory()` a separate class, with specific print method that prints more nicely. For example, list of files is not printed anymore, thus also changed the default of `add_files` in `get_inventory()` to `TRUE` instead of `FALSE`.
 - added workaround for "months since" calendar, which does not work well with ncdf4.helpers
+- added further check in `check_inventory()` to test if rcp scenarios have a corresponding historical run
 
 ---